save tif fix

+++ wsipack/wsi/save_tif.py
@@ -57,7 +57,7 @@ def main():
     parser = argparse.ArgumentParser(description="Create tissue masks")
     parser.add_argument("--wsi", help="slide path or directory containing slides", required=True)
     parser.add_argument("--out_dir", help="Output directory for the tissue masks", required=True)
-    parser.add_argument("--spacing", type=int, help="spacing in um/px (if not level)", required=False)
+    parser.add_argument("--spacing", type=float, help="spacing in um/px (if not level)", required=False)
     parser.add_argument("--level", type=int, help="level (if not spacing)", required=False)
     parser.add_argument("--suffix", default="tif,tiff,svs,ndpi,mrxs,dicom", help="suffix of slides in slide_dir", required=False)
     parser.add_argument("--overwrite", action="store_true", help="Overwrite existing files")
